feat(repo/web/api) mirroing mouse pointer so user can see who else are working in the board
