+++ apps/api/.gitignore
@@ -1,3 +1,5 @@
 node_modules/
 .build/*
 .env.dev
+key.pem
+cert.pem
