+++ apps/api/package.json
@@ -8,7 +8,7 @@
         "test": "echo \"Error: no test specified\" && exit 1",
         "dev": "nodemon",
         "build": "tsc",
-        "start": "npm run build && node .build/server.js"
+        "start": "npm run build && NODE_ENV=production && node .build/server.js"
     },
     "keywords": [],
     "author": "",
@@ -27,6 +27,7 @@
         "@types/cors": "^2.8.17",
         "@types/express": "^4.17.21",
         "@types/node": "^20.11.17",
+        "cpy-cli": "^5.0.0",
         "nodemon": "^3.0.3",
         "typescript": "^5.3.3"
     }
