+++ apps/web/package.json
@@ -32,6 +32,7 @@
         "react-dom": "^18.2.0",
         "react-router-dom": "^6.22.0",
         "socket.io": "^4.7.4",
+        "socket.io-client": "^4.7.4",
         "tailwind-merge": "^2.2.1",
         "tailwindcss-animate": "^1.0.7",
         "vaul": "^0.9.0"
