+++ apps/web/src/lib/utils/socketConfig.ts
@@ -1,0 +1,6 @@
+import { io } from 'socket.io-client';
+
+const URL = process.env.NODE_ENV === 'production' ? '' : 'http://localhost:3000';
+
+//ts-ignore
+export const socket = io(URL);
